Rename daemon-grpc to grpc
