+++ android/app/build.gradle.kts
@@ -374,7 +374,7 @@ play {
 
 dependencies {
     implementation(projects.lib.common)
-    implementation(projects.lib.daemonGrpc)
+    implementation(projects.lib.grpc)
     implementation(projects.lib.endpoint)
     implementation(projects.lib.map)
     implementation(projects.lib.model)
