+++ android/app/src/main/kotlin/net/mullvad/mullvadvpn/viewmodel/MullvadAppViewModel.kt
@@ -19,8 +19,8 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.shareIn
 import kotlinx.coroutines.launch
 import net.mullvad.mullvadvpn.compose.util.BackstackObserver
-import net.mullvad.mullvadvpn.lib.daemon.grpc.GrpcConnectivityState
-import net.mullvad.mullvadvpn.lib.daemon.grpc.ManagementService
+import net.mullvad.mullvadvpn.lib.grpc.GrpcConnectivityState
+import net.mullvad.mullvadvpn.lib.grpc.ManagementService
 import net.mullvad.mullvadvpn.lib.repository.ConnectionProxy
 
 private val noServiceDestinations = listOf(SplashDestination, PrivacyDisclaimerDestination)
