+++ android/lib/billing/src/main/kotlin/net/mullvad/mullvadvpn/lib/billing/PlayPurchaseRepository.kt
@@ -1,6 +1,6 @@
 package net.mullvad.mullvadvpn.lib.billing
 
-import net.mullvad.mullvadvpn.lib.daemon.grpc.ManagementService
+import net.mullvad.mullvadvpn.lib.grpc.ManagementService
 import net.mullvad.mullvadvpn.lib.model.PlayPurchase
 
 class PlayPurchaseRepository(private val managementService: ManagementService) {
