+++ android/lib/grpc/src/main/AndroidManifest.xml
