+++ android/lib/grpc/src/main/kotlin/net/mullvad/mullvadvpn/lib/grpc/mapper/FromDomain.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.lib.daemon.grpc.mapper
+package net.mullvad.mullvadvpn.lib.grpc.mapper
 
 import mullvad_daemon.management_interface.ManagementInterface
 import net.mullvad.mullvadvpn.lib.model.ApiAccessMethod
