+++ android/tile/build.gradle.kts
@@ -34,7 +34,7 @@ android {
 
 dependencies {
     implementation(projects.lib.common)
-    implementation(projects.lib.daemonGrpc)
+    implementation(projects.lib.grpc)
     implementation(projects.lib.model)
     implementation(projects.lib.ui.resource)
     implementation(projects.lib.repository)
